1. 전자문서 Controller 추가 2. URL 매핑 Action 클래스 추가 3. 일반문서 작성 구현

+++ src/main/java/com/hta2405/unite/dto/Doc.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @ToString
 public class Doc {
-    private int docId;
+    private Long docId;
     private String docWriter;
     private String docType;
     private String docTitle;
